Fix 5 spots of bad code
