+++ docker-compose.yml
@@ -7,14 +7,3 @@ services:
       MYSQL_DATABASE: mydb
     ports:
       - "3306:3306"
-
-  web:
-    build: .
-    environment:
-      FLASK_APP: app.py
-      FLASK_ENV: development
-      DATABASE_URL: mysql+pymysql://root:mypassword@db:3306/mydb
-    ports:
-      - "5000:5000"
-    depends_on:
-      - db
