+++ init_mysql.py
@@ -1,0 +1,5 @@
+from app import app
+from models import db
+
+with app.app_context():
+    db.create_all()
