+++ requirements.txt
@@ -1,3 +1,4 @@
 flask
 flask-sqlalchemy
 flask_restful
+pymysql
